chore(core): token-diet phase 2-C — split mur-core/src/cmd/agent.rs
Pure code movement, behavior unchanged. Splits the 1690-line
cmd/agent.rs into a folder so every file satisfies the project's
800-line limit:

  cmd/agent/
    mod.rs       (155)  shared helpers + pub use re-exports
    lifecycle.rs (589)  create / list / status / stop / remove / rename
                        + helpers (validate_name / classify / inbox guard)
    comm.rs      (142)  send / card (+ ephemeral runtime + dial_rpc)
    service.rs   (117)  install-service (launchd plist + systemd unit)
    perm.rs      (236)  cmd_perm_* + warn_if_running
    mcp.rs       (192)  cmd_mcp_* + McpAddPin
    skill.rs     (101)  cmd_skill_* + resolve_skill_id
    prompt.rs     56    cmd_prompt_show/edit/set + prompt_path_for
    export.rs     81    cmd_export + export_bin + locate_runtime_manifest_dir
    stats.rs      82    cmd_stats + cmd_logs
    secret.rs     63    cmd_secret_set/delete/list

Shared helpers (resolve_mur_home / resolve_bin_dir / resolve_runtime_target /
write_atomic / pid_alive / load_profile_for_edit / save_profile /
refuse_if_running) live in mod.rs with pub(super) or pub(crate)
visibility so each submodule and `crate::agent_admin` can call them.

`prompt_path_for` is exposed pub(crate) (used by agent_admin/prompt.rs).
`McpAddPin` re-exported from mcp.rs.

`#[allow(unused_imports)]` annotates the public CLI dispatch
re-exports — rustc flags them as unused under the lib+bin compilation
split because the lib itself never references the bound names; main.rs
and agent_admin reach them via the `crate::cmd::agent::cmd_*` path.

cargo build / cargo test -p mur-core (1114 lib + integration suites all
green) / cargo clippy -p mur-core --lib clean.
