+++ mur-core/src/cmd/agent/mod.rs
@@ -1,0 +1,159 @@
+//! `mur agent` subcommand dispatchers — split per resource:
+//!   - lifecycle (create/list/status/stop/remove/rename)
+//!   - comm      (send/card)
+//!   - service   (install-service)
+//!   - stats     (stats/logs)
+//!   - export    (cmd_export)
+//!   - perm      (cmd_perm_*)
+//!   - mcp       (cmd_mcp_*, McpAddPin)
+//!   - skill     (cmd_skill_*)
+//!   - prompt    (cmd_prompt_*)
+//!   - secret    (cmd_secret_*)
+//!
+//! This module re-exports the public entry points so `crate::cmd::agent::*`
+//! continues to resolve from main.rs and `crate::agent_admin`.
+//!
+//! P0a: just `create`.
+// TODO(Q-B): `mur agent rekey <name>` — regenerate identity keypair and
+// re-register with commander. Blocked on user decision in spec § 13.
+// If accepted, keep UUID stable; only rotate pubkey + notify peers.
+
+use std::fs;
+use std::path::{Path, PathBuf};
+
+use anyhow::{Context, Result, anyhow, bail};
+use mur_common::{AgentProfile as _AgentProfile, LockFile};
+
+mod comm;
+mod export;
+mod lifecycle;
+mod mcp;
+mod perm;
+mod prompt;
+mod secret;
+mod service;
+mod skill;
+mod stats;
+
+// `pub use` re-exports for the public CLI dispatch API. The lib crate doesn't
+// reference these names internally; consumers (main.rs, agent_admin) reach them
+// via `crate::cmd::agent::cmd_*`. Rustc still flags them as `unused_imports`
+// under the lib+bin compilation split — silenced here.
+#[allow(unused_imports)]
+pub use comm::{cmd_card, cmd_send};
+#[allow(unused_imports)]
+pub use export::cmd_export;
+#[allow(unused_imports)]
+pub use lifecycle::{cmd_create, cmd_list, cmd_remove, cmd_rename, cmd_status, cmd_stop};
+#[allow(unused_imports)]
+pub use mcp::{McpAddPin, cmd_mcp_add, cmd_mcp_list, cmd_mcp_remove, cmd_mcp_rename};
+#[allow(unused_imports)]
+pub use perm::{
+    cmd_perm_allow_host, cmd_perm_allow_read, cmd_perm_allow_spawn, cmd_perm_allow_write,
+    cmd_perm_deny_host, cmd_perm_deny_path, cmd_perm_deny_spawn, cmd_perm_list_hosts,
+    cmd_perm_set_limit, cmd_perm_set_mode, cmd_perm_show,
+};
+#[allow(unused_imports)]
+pub use prompt::{cmd_prompt_edit, cmd_prompt_set, cmd_prompt_show};
+#[allow(unused_imports)]
+pub(crate) use prompt::prompt_path_for;
+#[allow(unused_imports)]
+pub use secret::{cmd_secret_delete, cmd_secret_list, cmd_secret_set};
+#[allow(unused_imports)]
+pub use service::cmd_install_service;
+#[allow(unused_imports)]
+pub use skill::{cmd_skill_add, cmd_skill_list, cmd_skill_remove, cmd_skill_show};
+#[allow(unused_imports)]
+pub use stats::{cmd_logs, cmd_stats};
+
+// ─── Shared helpers used across submodules ─────────────────────────
+
+pub(crate) fn resolve_mur_home() -> Result<PathBuf> {
+    if let Some(v) = std::env::var_os("MUR_HOME") {
+        return Ok(PathBuf::from(v));
+    }
+    Ok(dirs::home_dir()
+        .ok_or_else(|| anyhow!("no home dir"))?
+        .join(".mur"))
+}
+
+pub(super) fn resolve_bin_dir() -> Result<PathBuf> {
+    if let Some(v) = std::env::var_os("MUR_AGENT_BIN_DIR") {
+        return Ok(PathBuf::from(v));
+    }
+    if let Some(home) = dirs::home_dir() {
+        return Ok(home.join(".local/bin"));
+    }
+    bail!("cannot resolve bin dir")
+}
+
+pub(super) fn resolve_runtime_target() -> PathBuf {
+    if let Some(v) = std::env::var_os("MUR_AGENT_RUNTIME_BIN") {
+        return PathBuf::from(v);
+    }
+    let runtime_filename = if cfg!(windows) {
+        "mur-agent-runtime.exe"
+    } else {
+        "mur-agent-runtime"
+    };
+    if let Ok(exe) = std::env::current_exe()
+        && let Some(dir) = exe.parent()
+    {
+        let candidate = dir.join(runtime_filename);
+        if candidate.exists() {
+            return candidate;
+        }
+    }
+    PathBuf::from(runtime_filename)
+}
+
+pub(super) fn write_atomic(path: &Path, bytes: &[u8]) -> Result<()> {
+    let tmp = path.with_extension(format!(
+        "{}.tmp",
+        path.extension().and_then(|e| e.to_str()).unwrap_or("tmp")
+    ));
+    fs::write(&tmp, bytes).with_context(|| format!("write {}", tmp.display()))?;
+    fs::rename(&tmp, path)
+        .with_context(|| format!("rename {} -> {}", tmp.display(), path.display()))?;
+    Ok(())
+}
+
+/// Liveness probe for a pid. Delegates to the canonical implementation in
+/// `mur_common::lock_file` so stop/remove/rename guards share the same logic
+/// as classify().
+pub(super) fn pid_alive(pid: u32) -> bool {
+    mur_common::lock_file::pid_alive(pid)
+}
+
+pub(crate) fn load_profile_for_edit(name: &str) -> Result<(PathBuf, _AgentProfile)> {
+    let mur_home = resolve_mur_home()?;
+    let path = mur_home.join("agents").join(name).join("profile.yaml");
+    if !path.exists() {
+        bail!("agent '{name}' not found");
+    }
+    let yaml = fs::read_to_string(&path).with_context(|| format!("read {}", path.display()))?;
+    let profile: _AgentProfile =
+        serde_yaml_ng::from_str(&yaml).with_context(|| format!("parse {}", path.display()))?;
+    Ok((path, profile))
+}
+
+pub(crate) fn save_profile(path: &Path, profile: &mut _AgentProfile) -> Result<()> {
+    profile.updated_at = chrono::Utc::now().to_rfc3339();
+    let yaml = serde_yaml_ng::to_string(profile).context("serialize profile.yaml")?;
+    write_atomic(path, yaml.as_bytes())
+}
+
+/// Used by stop/remove/rename to refuse mutation when the supervisor is alive.
+pub(super) fn refuse_if_running(agent_home: &Path, name: &str) -> Result<()> {
+    let lock_path = agent_home.join("running.lock");
+    if !lock_path.exists() {
+        return Ok(());
+    }
+    let bytes = fs::read(&lock_path).with_context(|| format!("read {}", lock_path.display()))?;
+    if let Ok(lock) = serde_json::from_slice::<LockFile>(&bytes)
+        && pid_alive(lock.pid)
+    {
+        bail!("agent '{name}' is running; stop it first");
+    }
+    Ok(())
+}
